Минимально работающая версия с простейшим полем.

+++ particle_random_walk/CPP/misc.cpp
@@ -6,7 +6,7 @@ int num_from_distribution(float* distribution, size_t length)
 	tmp_dist[0] = distribution[0];
 	for (int i = 1; i < length; ++i)
 		tmp_dist[i] = tmp_dist[i - 1] + distribution[i];
-	float generated = (double)rand() / double(RAND_MAX);
+	float generated = (float)rand() / (float)RAND_MAX;
 	int returned = 0;
 	for (int i = 0; i < length; ++i)
 		if (generated < tmp_dist[i])
